requirements.txt
annotated-types==0.7.0
anyio==4.5.2
boto3==1.35.77
botocore==1.35.77
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
numpy==1.24.4
pillow==10.4.0
pydantic==2.10.3
pydantic-core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
s3transfer==0.10.4
simplekml==1.3.6
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing-extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
